partial fix for PR #722 Work Item 1
This fixes #542
